# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: CallScope API
  version: 1.0.0
  description: |
    Session, ingest, review, evaluation and governance endpoints. Public endpoints are marked;
    all others require a service bearer token.

    Implemented: sessions/events (T-M1-07); review list/detail/audio/labels, labelled export,
    eval runs/compare, and model inventory (T-M4-02). Signed audio URLs expire in 5 minutes
    (HMAC query token). `/metrics` and `/metrics/eval` (eval FileEvalStore gauges, T-M4-05)
    are Prometheus scrape targets. T-M5-01: GET/POST /v1/models is seeded from the file-backed
    ModelStackRegistry (M0 backfill) and lists licence, status, intended use; transition gates
    land in T-M5-04. T-M5-04: lifecycle gates on POST /v1/models/{id}/transition return 409 with
    an ``unmet`` gate list when reports/cards/risks/monitoring/rollback are incomplete.
servers:
  - url: https://callscope.example.com
security:
  - serviceToken: []
paths:
...
  /v1/models:
    get:
      summary: Model inventory
      parameters:
        - { name: component, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string } }
      responses:
        '200':
          description: Model versions
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ModelVersion' }
    post:
      summary: Register a model version (status starts as candidate)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ModelVersionCreate' }
      responses:
        '201':
          description: Registered
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ModelVersion' }
  /v1/models/{modelVersionId}/transition:
    post:
      summary: Lifecycle transition (gates enforced server-side)
      parameters: [ { name: modelVersionId, in: path, required: true, schema: { type: string, format: uuid } } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [to]
              properties:
                to: { type: string, enum: [validated, production, retired, rejected] }
                report_id: { type: string, format: uuid }
                report_passed: { type: boolean }
                monitoring_on: { type: boolean, default: true }
                rollback_stack_id: { type: string }
                rationale: { type: string }
      responses:
        '200':
          description: Transitioned
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ModelVersion' }
        '409':
          description: Gate not satisfied
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
...
components:
  securitySchemes:
    serviceToken:
      type: http
      scheme: bearer
...
  schemas:
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        code: { type: string }
...
    ModelVersionCreate:
      type: object
      required: [component, name, revision, owner]
      properties:
        component: { type: string, enum: [asr, tts, llm, vad, turn_detector, judge] }
        name: { type: string }
        base_model: { type: string }
        revision: { type: string }
        license: { type: string }
        artifact_uri: { type: string }
        config: { type: object }
        intended_use: { type: string }
        out_of_scope_use: { type: string }
        owner: { type: string }
        mlflow_run_id: { type: string }
    ModelVersion:
      allOf:
        - $ref: '#/components/schemas/ModelVersionCreate'
        - type: object
          properties:
            model_version_id: { type: string, format: uuid }
            status: { type: string, enum: [candidate, validated, production, retired, rejected] }
            created_at: { type: string, format: date-time }
...